refactor(styleboxflat): 将边框宽度相关类型从int改为float
调整了StyleBoxFlat的边框宽度接口、成员变量和文档，支持浮点精度的边框宽度设置，同时更新了属性提示的步进值为0.5

+++ scene/resources/style_box_flat.h
@@ -65,11 +65,11 @@ public:
 	void set_border_color(const Color &p_color);
 	Color get_border_color() const;
 
-	void set_border_width_all(int p_size);
-	int get_border_width_min() const;
+	void set_border_width_all(float p_size);
+	float get_border_width_min() const;
 
-	void set_border_width(Side p_side, int p_width);
-	int get_border_width(Side p_side) const;
+	void set_border_width(Side p_side, float p_width);
+	float get_border_width(Side p_side) const;
 
 	void set_border_blend(bool p_blend);
 	bool get_border_blend() const;
